+++ GRDB/Legacy/Fixits-0.109.0.swift
@@ -1,0 +1,4 @@
+extension Request {
+    @available(*, unavailable, renamed:"fetching(_:)")
+    public func bound<T>(to type: T.Type) -> AnyTypedRequest<T> { preconditionFailure() }
+}